refactor(engines): extract the decision seam from strategy.py
Move per-symbol decision production (_process_symbol) and decision routing
(_handle_decision: proposal-only mode, BUY/SELL placement, CLOSE reduce-only
exits) plus the proposal-mapping helpers into engines/decision_flow.py,
mirroring the cycle_runner seam (module functions taking the engine).
strategy.py keeps two thin delegates so call sites and tests are unchanged;
the proposal helpers had no external callers and moved outright.

strategy.py shrinks 2852 -> 2441 lines; decision_flow.py is 443.

Stacked on test/engine-fixture-depin: the de-pinned real-flow order tests
exercise the extracted path end to end over the production guard stack.
